+++ src/Initializer.php
@@ -35,6 +35,7 @@ class Initializer implements InitializerContracts
     'eventlog',
     'exception',
     'seeds',
+    'helpers',
   ];
 
   protected $tasks = [];
@@ -102,6 +103,7 @@ class Initializer implements InitializerContracts
       yield $this->composer->add('primalbase/view-builder', 'dev-master');
       yield $this->composer->add('eyewill/tucle-builder', 'dev-master');
       yield $this->composer->add('bugsnag/bugsnag-laravel', '^2.0');
+      yield $this->composer->addAutoload('files', 'app/Http/helpers.php');
     }
 
     if (in_array('config', $this->tasks))
@@ -117,6 +119,14 @@ class Initializer implements InitializerContracts
       yield $this->makeFromStub(
         __DIR__.'/../files/config/.env.stub',
         $this->basePath.'/.env.local'
+      );
+    }
+
+    if (in_array('helpers', $this->tasks))
+    {
+      yield $this->makeFromStub(
+        __DIR__.'/../files/helpers.stub',
+        $this->app['path'].'/helpers.php'
       );
     }
 
@@ -157,7 +167,15 @@ class Initializer implements InitializerContracts
 
     if (in_array('routes', $this->tasks))
     {
-      yield $this->makeHttpRoutes();
+      yield $this->makeFromStub(
+        __DIR__ . '/../files/Http/routes.stub',
+        $this->app['path'].'/Http/routes.php'
+      );
+
+      yield $this->makeFromStub(
+        __DIR__ . '/../files/Http/Frontend/routes.stub',
+        $this->app['path'].'/Http/Frontend/routes.php'
+      );
     }
 
     if (in_array('providers', $this->tasks))
@@ -329,21 +347,6 @@ class Initializer implements InitializerContracts
     $this->app['files']->copy(__DIR__.'/../files/gulpfile.js', $path);
 
     return $path.' copied.';
-  }
-
-  public function makeHttpRoutes()
-  {
-    $filePath = $this->app['path'].'/Http/routes.php';
-    if (!$this->force && $this->app['files']->exists($filePath)) {
-      return $filePath . ' already exists.';
-    }
-
-    $this->app['files']->makeDirectory(dirname($filePath), 02755, true, true);
-    $templatePath = __DIR__.'/../files/routes.stub';
-    $template = $this->app['files']->get($templatePath);
-    $this->app['files']->put($filePath, $template);
-
-    return $filePath.' generated.';
   }
 
   public function makeEventLogMigrationFile()